Move custom types for v3 and v4 to algoliasearch.ts
