+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from './instantsearch';
 export * from './connector';
 export * from './widget';
 export * from './insights';
+export * from './algoliasearch';
