+++ src/widgets/infinite-hits/__tests__/infinite-hits-test.ts
@@ -1,6 +1,6 @@
 import { render as preactRender } from 'preact';
 import algoliasearchHelper from 'algoliasearch-helper';
-import { Client } from '../../../types';
+import { AlgoliaSearchClient as Client } from '../../../types';
 import infiniteHits from '../infinite-hits';
 import { castToJestMock } from '../../../../test/utils/castToJestMock';
